refactor: LottoMachine 당첨금액 LottoCalculator 로 역할 위임
