+++ __tests__/LottoMachine.test.js
@@ -1,52 +1,25 @@
 import LottoMachine from '../src/js/domain/LottoMachine';
 import LottoTicket from '../src/js/domain/LottoTicket';
+import { ERROR_MESSAGE } from '../src/js/constants';
 
 const context = describe;
-const INVALID_WINNING_AMOUNTS = [
-  [[]],
-  [[1, 2, 3, 4, 5]],
-  [[1, 2, 3, 4, 5, 6, 7]],
-];
-
-describe('당첨 금액 설정 테스트', () => {
-  context.each(INVALID_WINNING_AMOUNTS)(
-    '%s와 같이 유효하지 않은 당첨 금액이 주어질 때',
-    (winningAmount) => {
-      it('TypeError를 Throw 한다.', () => {
-        expect(() => new LottoMachine(winningAmount)).toThrow();
-      });
-    }
-  );
-});
 
 describe('로또 발권 테스트', () => {
   context('로또금액 1_000원 보다 적은 800을 지불할 때', () => {
-    it(`"${LottoMachine.NOT_ENOUGH_MONEY}" 에러를 Throw한다.`, () => {
-      const lottoMachine = new LottoMachine();
-
-      expect(() => lottoMachine.sellAutoLottoTicket(800)).toThrow();
+    it(`"${ERROR_MESSAGE.NOT_ENOUGH_MONEY}" 에러를 Throw한다.`, () => {
+      expect(() => LottoMachine.sellAutoLottoTicket(800)).toThrow();
     });
   });
 
   context('로또 금액 1_000원 보다 많은 2_000원 을 지불할 때', () => {
     it('lottoTickets의 길이는 2이다.', () => {
-      // given
-      const lottoMachine = new LottoMachine();
+      const lottoTickets = LottoMachine.sellAutoLottoTicket(2_000);
 
-      // when
-      const lottoTickets = lottoMachine.sellAutoLottoTicket(2_000);
-
-      // then
       expect(lottoTickets).toHaveLength(2);
     });
     it('lottoTickets은 모두 LottoTicket 인스턴스를 가져야 한다. 기대값: true', () => {
-      // given
-      const lottoMachine = new LottoMachine();
+      const lottoTickets = LottoMachine.sellAutoLottoTicket(2_000);
 
-      // when
-      const lottoTickets = lottoMachine.sellAutoLottoTicket(2_000);
-
-      // then
       expect(
         lottoTickets.every((lotto) => lotto instanceof LottoTicket)
       ).toBeTruthy();
@@ -57,9 +30,7 @@ describe('로또 발권 테스트', () => {
     'sellAutoLottoTicket() 매개변수로 %s가 주어질 때',
     (invalidValue) => {
       it('TypeError를 Throw한다.', () => {
-        const lottoMachine = new LottoMachine();
-
-        expect(() => lottoMachine.sellAutoLottoTicket(invalidValue)).toThrow();
+        expect(() => LottoMachine.sellAutoLottoTicket(invalidValue)).toThrow();
       });
     }
   );
