+++ src/js/constants/index.js
@@ -20,4 +20,5 @@ export const ERROR_MESSAGE = {
   REQUIRE_LOTTO_INSTANCE: '로또가 아닙니다.',
   MIN_LENGTH: '최소 길이를 만족하지 않습니다.',
   MAX_LENGTH: '최대 길이를 초과하였습니다.',
+  NOT_ENOUGH_MONEY: '돈이 부족합니다.',
 };
